requirements.txt
rumps>=0.4.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.2
yt-dlp>=2024.1.1
pyobjc-framework-Cocoa>=10.0
mcp[cli]>=1.2
